requirements.txt
asgiref==3.7.2
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
distlib==0.3.6
Django==4.2.5
et-xmlfile==1.1.0
filelock==3.9.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lml==0.1.0
MarkupSafe==2.1.3
mysql-client==0.0.1
mysqlclient==2.2.0
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
platformdirs==3.0.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyFirmata==1.1.0
PyJWT==2.8.0
pyserial==3.5
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
sqlparse==0.4.4
texttable==1.7.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.19.0
Werkzeug==2.3.7
